+++ src/it/test-projects/creedengo-java-plugin-test-project/src/main/java/org/greencodeinitiative/creedengo/java/checks/UseEveryColumnQueriedLitteralQueryNonCompliant.java
@@ -24,10 +24,10 @@ import java.sql.SQLException;
 import java.sql.Statement;
 
 /**
- * In this test case, the ResultSet is passed through a method
- * All Fields are accessed, so no issue is raised
+ * In this test case, the query is a litteral string directly inserted in the method
+ * One field is not accesed, so an issue is raised
  */
-public class UseMethodCompliant {
+public class UseEveryColumnQueriedLitteralQueryNonCompliant {
 
 	private static final String DB_URL = "jdbc:mysql://localhost/TEST";
 	private static final String USER = "guest";
@@ -37,21 +37,18 @@ public class UseMethodCompliant {
 
 		try (Connection conn = DriverManager.getConnection(DB_URL, USER, PASS);
 				Statement stmt = conn.createStatement();
-				ResultSet rs = stmt.executeQuery("SELECT id, first, last, age FROM Registration");) {
-			extractGet(rs);
+				ResultSet rs = stmt.executeQuery("SELECT id, first, last, age FROM Registration");) { // Noncompliant {{Avoid querying SQL columns that are not used}}
+			while (rs.next()) {
+				// Display values
+				System.out.print("ID: " + rs.getInt("id"));
+				System.out.print(", First: " + rs.getString("first"));
+				System.out.println(", Last: " + rs.getString("last"));
+			}
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
 
 	}
-
-	private void extractGet(ResultSet rs) throws SQLException {
-		while (rs.next()) {
-			// Display values
-			System.out.print("ID: " + rs.getInt("id"));
-			System.out.print(", Age: " + rs.getInt("age"));
-			System.out.print(", First: " + rs.getString("first"));
-			System.out.println(", Last: " + rs.getString("last"));
-		}
-	}
+	
+	
 }