+++ src/it/test-projects/creedengo-java-plugin-test-project/src/main/java/org/greencodeinitiative/creedengo/java/checks/UseEveryColumnQueriedUseColumnIdsAndNameAttributesNonCompliant.java
@@ -24,24 +24,26 @@ import java.sql.SQLException;
 import java.sql.Statement;
 
 /**
- * In this test case, the query is a litteral string directly inserted in the method
+ * In this test case, columns are accesed by IDs and names, some of them being in final variables
  * One field is not accesed, so an issue is raised
  */
-public class LitteralQueryNonCompliant {
+public class UseEveryColumnQueriedUseColumnIdsAndNameAttributesNonCompliant {
 
 	private static final String DB_URL = "jdbc:mysql://localhost/TEST";
 	private static final String USER = "guest";
 	private static final String PASS = "guest123";
+	private static final String QUERY = "SELECT id, first, last, age FROM Registration";
+	private static final String ID = "id";
 
 	public void callJdbc() {
 
 		try (Connection conn = DriverManager.getConnection(DB_URL, USER, PASS);
 				Statement stmt = conn.createStatement();
-				ResultSet rs = stmt.executeQuery("SELECT id, first, last, age FROM Registration");) { // Noncompliant {{Avoid querying SQL columns that are not used}}
+				ResultSet rs = stmt.executeQuery(QUERY);) { // Noncompliant {{Avoid querying SQL columns that are not used}}
 			while (rs.next()) {
 				// Display values
-				System.out.print("ID: " + rs.getInt("id"));
-				System.out.print(", First: " + rs.getString("first"));
+				System.out.print("ID: " + rs.getInt(ID));
+				System.out.print(", First: " + rs.getString(2));
 				System.out.println(", Last: " + rs.getString("last"));
 			}
 		} catch (SQLException e) {
@@ -49,6 +51,4 @@ public class LitteralQueryNonCompliant {
 		}
 
 	}
-	
-	
 }