+++ src/test/files/UseEveryColumnQueried/UseEveryColumnQueriedMultipleQueriesCompliant.java
@@ -24,31 +24,46 @@ import java.sql.SQLException;
 import java.sql.Statement;
 
 /**
- * In this test case, the query is a litteral string directly inserted in the method
- * One field is not accesed, so an issue is raised
+ * In this test case, multiple queries are done using the same Statement Object.
+ * All Fields are accessed, so no issue is raised
  */
-public class LitteralQueryNonCompliant {
+public class UseEveryColumnQueriedMultipleQueriesCompliant {
 
 	private static final String DB_URL = "jdbc:mysql://localhost/TEST";
 	private static final String USER = "guest";
 	private static final String PASS = "guest123";
-
+	private static final String QUERY = "SELECT id, first, last, age FROM Registration"; 
+	private static final String QUERY2 = "SELECT id, first, last FROM Registration2";
+	
+	
 	public void callJdbc() {
 
 		try (Connection conn = DriverManager.getConnection(DB_URL, USER, PASS);
 				Statement stmt = conn.createStatement();
-				ResultSet rs = stmt.executeQuery("SELECT id, first, last, age FROM Registration");) { // Noncompliant {{Avoid querying SQL columns that are not used}}
+				) {
+			
+			ResultSet rs = stmt.executeQuery(QUERY);
+			while (rs.next()) {
+				// Display values
+				System.out.print("Age: " + rs.getInt("age"));
+				System.out.print("ID: " + rs.getInt("id"));
+				System.out.print(", First: " + rs.getString("first"));
+				System.out.println(", Last: " + rs.getString("last"));
+			}
+			rs = stmt.executeQuery(QUERY2);
+			
+			
 			while (rs.next()) {
 				// Display values
 				System.out.print("ID: " + rs.getInt("id"));
 				System.out.print(", First: " + rs.getString("first"));
 				System.out.println(", Last: " + rs.getString("last"));
 			}
+			
+			
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
 
 	}
-	
-	
 }