+++ src/test/files/UseEveryColumnQueried/UseEveryColumnQueriedUseMethodCompliant.java
@@ -24,31 +24,34 @@ import java.sql.SQLException;
 import java.sql.Statement;
 
 /**
- * In this test case, columns are accesed by IDs and names, some of them being in final variables
- * One field is not accesed, so an issue is raised
+ * In this test case, the ResultSet is passed through a method
+ * All Fields are accessed, so no issue is raised
  */
-public class UseColumnIdsAndNameAttributesNonCompliant {
+public class UseEveryColumnQueriedUseMethodCompliant {
 
 	private static final String DB_URL = "jdbc:mysql://localhost/TEST";
 	private static final String USER = "guest";
 	private static final String PASS = "guest123";
-	private static final String QUERY = "SELECT id, first, last, age FROM Registration";
-	private static final String ID = "id";
 
 	public void callJdbc() {
 
 		try (Connection conn = DriverManager.getConnection(DB_URL, USER, PASS);
 				Statement stmt = conn.createStatement();
-				ResultSet rs = stmt.executeQuery(QUERY);) { // Noncompliant {{Avoid querying SQL columns that are not used}}
-			while (rs.next()) {
-				// Display values
-				System.out.print("ID: " + rs.getInt(ID));
-				System.out.print(", First: " + rs.getString(2));
-				System.out.println(", Last: " + rs.getString("last"));
-			}
+				ResultSet rs = stmt.executeQuery("SELECT id, first, last, age FROM Registration");) {
+			extractGet(rs);
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
 
+	}
+
+	private void extractGet(ResultSet rs) throws SQLException {
+		while (rs.next()) {
+			// Display values
+			System.out.print("ID: " + rs.getInt("id"));
+			System.out.print(", Age: " + rs.getInt("age"));
+			System.out.print(", First: " + rs.getString("first"));
+			System.out.println(", Last: " + rs.getString("last"));
+		}
 	}
 }